Added guards for data-payloads
Someone reported that Moblin (who uses this library) crashed whenever they connected their cooler. The stacktrace from Erik claims the issue is in the parseMessages call. Only thing i can see that might be an issue there is if the length of the data is not sufficiently long enough for the bits we want to read.